Fixed social header.
Title now working in all cases. Also stopped using deprecated wp_title.
URL working in all cases.

Image picks the image of the first post. I think that's ok.

Description still empty.

Cleaned up functions.php a little
