+++ theme/Makefile
@@ -27,7 +27,8 @@ phpFiles = $(standardTemplates) \
 	   card.php \
 	   card-gallery.php \
 	   content-none.php \
-	   comments.php
+	   comments.php \
+	   fancy-photo.php
 styleSheets = style.css 
 files = $(jsFiles) $(phpFiles) $(styleSheets) screenshot.png
 images = arrows.png feedicons-standard/feed-icon-14x14.png lightpaperfibers.png toolbar_find.png green_cup.png seamlesstexture15_500.jpg pageTurn.png bookBinding.png
